iwidjango-master/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
httpie==0.9.9
idna==2.6
Markdown==2.6.8
MySQL-python==1.2.4
numpy==1.13.0
psycopg2==2.7.1
pygame==1.9.3
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.0
